requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
asyncpg==0.29.0
psycopg2-binary==2.9.9
jinja2==3.1.2
python-dotenv==1.0.0
ib_async==2.0.1
aiofiles==23.2.1
python-multipart
httpx==0.25.2
